tc-build: Allow user to specify whatever CFLAGS they want
As it turns out, -march=native is not necessarily the best optimization
flag...

https://github.com/ClangBuiltLinux/tc-build/pull/55

The use of AVX512 instructions causes certain Intel CPUs to downclock,
which can ruin the performance of certain workloads, such as compiling
the Linux kernel.

Get rid of the -march=native flag to avoid these performance slow downs.
Instead of leaving users with no way to customize their CFLAGS, add a
--cflags parameter so the user can supply whatever flags they want; at
that point, it is on them to measure the performance they gain or lose
from those flags.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>